# src/frontend/sqeez/src/api/api.yaml
openapi: 3.1.1
info:
  title: Sqeez.Api | v1
  version: 1.0.0
servers:
  - url: http://localhost:5000/
paths:
...
  /api/media-assets:
    get:
      tags:
        - MediaAssets
      parameters:
        - name: SearchTerm
          in: query
          schema:
            type: string
        - name: MimeType
          in: query
          schema:
            $ref: "#/components/schemas/MediaType"
        - name: IsPrivate
          in: query
          schema:
            type: boolean
        - name: OwnerId
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        - name: PageNumber
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: PageSize
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/PagedResponseOfMediaAssetDto"
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfMediaAssetDto"
            text/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfMediaAssetDto"
    post:
      tags:
        - MediaAssets
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaAssetDto"
          text/json:
            schema:
              $ref: "#/components/schemas/CreateMediaAssetDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/CreateMediaAssetDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
            text/json:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
  /api/media-assets/{id}:
    get:
      tags:
        - MediaAssets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
            text/json:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
    patch:
      tags:
        - MediaAssets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchMediaAssetDto"
          text/json:
            schema:
              $ref: "#/components/schemas/PatchMediaAssetDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/PatchMediaAssetDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
            text/json:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
    delete:
      tags:
        - MediaAssets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
  /api/media-assets/upload:
    post:
      tags:
        - MediaAssets
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                File:
                  $ref: "#/components/schemas/IFormFile"
                OwnerId:
                  pattern: ^-?(?:0|[1-9]\d*)$
                  type:
                    - integer
                    - string
                  format: int64
                IsPrivate:
                  type: boolean
                Description:
                  type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
            text/json:
              schema:
                $ref: "#/components/schemas/MediaAssetDto"
  /api/media-assets/{id}/file:
    get:
      tags:
        - MediaAssets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/PhysicalFileResult"
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicalFileResult"
            text/json:
              schema:
                $ref: "#/components/schemas/PhysicalFileResult"
...
components:
  schemas:
...
    CreateMediaAssetDto:
      required:
        - locationUrl
        - mimeType
        - isPrivate
        - ownerId
      type: object
      properties:
        locationUrl:
          type: string
        mimeType:
          $ref: "#/components/schemas/MediaType"
        isPrivate:
          type: boolean
        ownerId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        description:
          type:
            - "null"
            - string
...
    EntityTagHeaderValue:
      type: object
      properties:
        tag:
          $ref: "#/components/schemas/StringSegment"
        isWeak:
          type: boolean
...
    IFormFile:
      type: string
      format: binary
...
    MediaAssetDto:
      required:
        - id
        - locationUrl
        - mimeType
        - isPrivate
        - description
        - ownerId
        - ownerUsername
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        locationUrl:
          type: string
        mimeType:
          $ref: "#/components/schemas/MediaType"
        isPrivate:
          type: boolean
        description:
          type:
            - "null"
            - string
        ownerId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        ownerUsername:
          type:
            - "null"
            - string
    MediaType:
      enum:
        - Image
        - Video
        - Audio
        - Document
...
    PagedResponseOfMediaAssetDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MediaAssetDto"
        totalCount:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageNumber:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
...
    PatchMediaAssetDto:
      type: object
      properties:
        locationUrl:
          type:
            - "null"
            - string
        mimeType:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/MediaType"
        isPrivate:
          type:
            - "null"
            - boolean
        description:
          type:
            - "null"
            - string
...
    PhysicalFileResult:
      type: object
      properties:
        fileName:
          type: string
        contentType:
          type:
            - "null"
            - string
        fileDownloadName:
          type:
            - "null"
            - string
        lastModified:
          type:
            - "null"
            - string
          format: date-time
        entityTag:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/EntityTagHeaderValue"
        enableRangeProcessing:
          type: boolean
...
    StringSegment:
      type: object
      properties:
        buffer:
          type:
            - "null"
            - string
        offset:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        length:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        value:
          type:
            - "null"
            - string
        hasValue:
          type: boolean
...